+++ mtg_proxy_printer/page_scene/page_scene.py
@@ -448,17 +448,32 @@ class PageScene(QGraphicsScene):
             self._update_page_number_text()
 
     def on_rows_moved(self, parent: QModelIndex, start: int, end: int, destination: QModelIndex, row: int):
-        if parent.isValid() and parent.row() == self.selected_page.row():
+        source_page_row = parent.row()
+        current_page_row = self.selected_page.row()
+        destination_page_row = destination.row()
+        if not parent.isValid():
+            # Moved pages around. Needs to update the current page text
+            self._update_page_number_text()
+            return
+        # Parent is valid, thus start:end+1 point to cards on the page source_page_row.
+        if source_page_row != current_page_row == destination_page_row:
+            # Cards moved onto the current page are treated as if they were added
+            logger.debug("Cards moved onto the currently shown page, calling card insertion handler.")
+            self.on_rows_inserted(destination, row, row + end - start)
+        elif source_page_row == current_page_row != destination_page_row:
             # Cards moved away are treated as if they were deleted
             logger.debug("Cards moved away from the currently shown page, calling card removal handler.")
             self.on_rows_removed(parent, start, end)
-        if destination.isValid() and destination.row() == self.selected_page.row():
-            # Moved in cards are treated as if they were added
-            logger.debug("Cards moved onto the currently shown page, calling card insertion handler.")
-            self.on_rows_inserted(destination, row, row+end-start)
-        if not parent.isValid() and not destination.isValid():
-            # Moved pages around. Needs to update the current page text
-            self._update_page_number_text()
+        elif source_page_row == current_page_row == destination_page_row:
+            logger.debug("Cards moved within the current page, reordering them")
+            card_items = self.card_items
+            # Dropping at the end has row pointing one item out of bounds
+            target_item = card_items[min(row, len(card_items)-1)]
+            for item in card_items[start:end + 1]:
+                item.stackBefore(target_item)
+            self.update_card_positions()
+        # Remaining cases are card moves happening "off-screen", so nothing has to be done on them.
+
 
     @functools.lru_cache(None)
     def _compute_position_for_image(self, index_row: int, page_type: PageType) -> QPointF: